Optimize the code compatibility of flink EngineConn, and optimize the pom parent of flink module.
