+++ linkis-engineconn-plugins/engineconn-plugins/flink/pom.xml
@@ -18,10 +18,9 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>linkis</artifactId>
+        <artifactId>linkis-engineconn-plugins</artifactId>
         <groupId>com.webank.wedatasphere.linkis</groupId>
         <version>1.0.2</version>
-        <relativePath>../../pom.xml</relativePath>
     </parent>
     <modelVersion>4.0.0</modelVersion>
