+++ linkis-engineconn-plugins/engineconn-plugins/flink/src/main/java/com/webank/wedatasphere/linkis/engineconnplugin/flink/client/sql/operation/OperationFactory.java
@@ -19,115 +19,15 @@
 package com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation;
 
 
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.CreateViewOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.DDLOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.DescribeTableOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.DropViewOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ExplainOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.InsertOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ResetOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.SelectOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.SetOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ShowCatalogsOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ShowCurrentCatalogOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ShowCurrentDatabaseOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ShowDatabasesOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ShowFunctionsOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ShowModulesOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ShowTablesOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.ShowViewsOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.UseCatalogOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.operation.impl.UseDatabaseOperation;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.utils.SqlCommandParser.SqlCommandCall;
+import com.webank.wedatasphere.linkis.engineconnplugin.flink.client.sql.parser.SqlCommandCall;
 import com.webank.wedatasphere.linkis.engineconnplugin.flink.context.FlinkEngineConnContext;
 import com.webank.wedatasphere.linkis.engineconnplugin.flink.exception.SqlParseException;
-import com.webank.wedatasphere.linkis.engineconnplugin.flink.ql.GrammarFactory;
 
-public class OperationFactory {
+public interface OperationFactory {
 
-	public static Operation createOperation(SqlCommandCall call, FlinkEngineConnContext context) throws SqlParseException {
-
-		Operation operation;
-		switch (call.command) {
-			case SELECT:
-				operation = new SelectOperation(context, call.operands[0]);
-				break;
-			case CREATE_VIEW:
-				operation = new CreateViewOperation(context, call.operands[0], call.operands[1]);
-				break;
-			case DROP_VIEW:
-				operation = new DropViewOperation(context, call.operands[0], Boolean.parseBoolean(call.operands[1]));
-				break;
-			case LINKIS_GRAMMAR:
-				operation = GrammarFactory.apply(call.operands[0], context);
-				break;
-			case CREATE_TABLE:
-			case DROP_TABLE:
-			case ALTER_TABLE:
-			case CREATE_DATABASE:
-			case DROP_DATABASE:
-			case ALTER_DATABASE:
-				operation = new DDLOperation(context, call.operands[0], call.command);
-				break;
-			case SET:
-				// list all properties
-				if (call.operands.length == 0) {
-					operation = new SetOperation(context);
-				} else {
-					// set a property
-					operation = new SetOperation(context, call.operands[0], call.operands[1]);
-				}
-				break;
-			case RESET:
-				if (call.operands.length > 0) {
-					throw new SqlParseException("Only RESET ALL is supported now");
-				}
-				operation = new ResetOperation(context);
-				break;
-			case USE_CATALOG:
-				operation = new UseCatalogOperation(context, call.operands[0]);
-				break;
-			case USE:
-				operation = new UseDatabaseOperation(context, call.operands[0]);
-				break;
-			case INSERT_INTO:
-			case INSERT_OVERWRITE:
-				operation = new InsertOperation(context, call.operands[0], call.operands[1]);
-				break;
-			case SHOW_MODULES:
-				operation = new ShowModulesOperation(context);
-				break;
-			case SHOW_CATALOGS:
-				operation = new ShowCatalogsOperation(context);
-				break;
-			case SHOW_CURRENT_CATALOG:
-				operation = new ShowCurrentCatalogOperation(context);
-				break;
-			case SHOW_DATABASES:
-				operation = new ShowDatabasesOperation(context);
-				break;
-			case SHOW_CURRENT_DATABASE:
-				operation = new ShowCurrentDatabaseOperation(context);
-				break;
-			case SHOW_TABLES:
-				operation = new ShowTablesOperation(context);
-				break;
-			case SHOW_VIEWS:
-				operation = new ShowViewsOperation(context);
-				break;
-			case SHOW_FUNCTIONS:
-				operation = new ShowFunctionsOperation(context);
-				break;
-			case DESCRIBE_TABLE:
-				operation = new DescribeTableOperation(context, call.operands[0]);
-				break;
-			case EXPLAIN:
-				operation = new ExplainOperation(context, call.operands[0]);
-				break;
-			default:
-				throw new SqlParseException("Unsupported command call " + call + ". This is a bug.");
-		}
-
-		return operation;
+	static OperationFactory getOperationFactory() {
+		return OperationFactoryImpl.getInstance();
 	}
+
+	Operation createOperation(SqlCommandCall call, FlinkEngineConnContext context) throws SqlParseException;
 }
